requirements.txt
python-dotenv
requests
telethon
asyncio
google-cloud-bigquery